+++ src/modules/synchro/components/HeadCard.module.scss
