+++ src/modules/synchro/components/ListCard.module.scss
