+++ src/modules/synchro/components/ListItem.module.scss
