+++ src/modules/synchro/components/SummaryCard.module.scss
