+++ src/modules/synchro/components/icons.tsx
